Fix arbitrage 500, ship class server-side filter, Devoid region, test updates
- Fix arbitrage Integer overflow: SUM(volume_remain) can exceed Integer.MAX_VALUE
  for high-volume items (e.g. Tritanium); changed volumeAvailable to Long in
  RegionMinPrice projection and ArbitrageOpportunityDto
- Move ship class filter server-side (capitalGroupName param) so rare classes
  like Titan/Supercarrier/Force Aux are filtered in DB, not on the client page
- Add Devoid region (10000036) to contract scanner, controller, and frontend
- Fix pruneExpiredContracts transaction error (same @Modifying fix as market scanner)
- Update ContractController/ContractRepository/tests for new capitalGroupName param

+++ backend/src/main/java/com/evemarket/backend/dto/ArbitrageOpportunityDto.java
@@ -15,7 +15,7 @@ public class ArbitrageOpportunityDto {
     private String     sellRegionName;
     private BigDecimal sellPrice;
     private Double     gapPercent;
-    private Integer    volumeAvailable;
+    private Long       volumeAvailable;
     private BigDecimal averagePrice;
     private boolean    alreadyListed;
 }
